requirements.txt
canvasapi
PyGithub